requirements.txt
python-telegram-bot>=22.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0

